build: Migrate to Poetry
Poetry is starting to be the standard used for the rest of Toolforge
infrastructure written in Python, and the lock file is a nice feature
for a fully containerized app. This commit migrates the app to use
Poetry. Dependencies are pinned to the currently running versions and a
follow-up will upgrade them.

Bug: T354751
Change-Id: If2385b4b44b5ffdd05be14c723e9564d5e40bbb9