jobs-api: introduce securityContext in the pod template
Introduce a stricter pod template by including an explicit securityContext to limit
what a pod can and cannot do.

This is part of the PSP migration. This will help us later when we start
enforcing that these values are always present for used-defined workloads.

Anyway, the kubernetes defaults for these values may or may not be what we
are expecting, so setting them explicitly should be fine even if not part of
the PSP migration.

This includes a new getent syscall everytime a job is create. Because of this,
we need to mount the nsswitch config from the host worker node. The nsswitch config
will point to sssd, which should be reachable via the sssd socket which is also
a hostpath mount. Eventually, the getent will hit the k8s worker node sssd daemon, then LDAP.

Bug: T362050
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>