api: allow protocol to be specified for ports
Currently `port` is configured implicitly as `TCP` in the `Job` and
explicitly as `TCP` in the `Service`.

This change adds `port_protocol` (enum `TCP`|`UDP`) which sets the
`containerPort` `protocol` explicitly on both the `Job` and `Service`.

Updating a `Job` specification to use `port_protocol: UDP`, rather than
the default of `TCP` will allow a `UDP` based service (`Job`) to function.

Bug: T400024\
Signed-off-by: David Caro <dcaro@wikimedia.org>