buildservice: strip `launcher` when returning the job
We add 'launcher' to the job command when creating it for
bulidservice-based jobs if it's not there. But we were not stripping
it off when returning them.

As we don't want users to have to specify the `launcher` bit
themselves, this patch removes it when returning the jobs, so commands
like `toolforge jobs dump` will not return it.

This also stripts it when comparing jobs, so if the user specified it,
it will still not be counted as a difference.

This will not be an issue once we store the original user command.

Signed-off-by: David Caro <dcaro@wikimedia.org>