api: add `include_unset` parameter to get_job and get_jobs
This allows to fetch the job with all the fields set, even if they
were not passed (default), and to get it also only with the fields
that were set on creation, so we can both show the user all the
currently operational field values (ex. jobs list, jobs show) or just
return the minimal parameters to create the job (ex. for jobs dump).

Bug: T402569
Signed-off-by: David Caro <dcaro@wikimedia.org>