CommonJob - Invert default mount option
The currently default of All is being applied to build pack based jobs,
as we want to allow users to specify mounting all in build pack based
jobs, it is easier to check the invariant.

New behaviour:
- If no `mount` option is specified or `mount` is specified as `NONE`
    `NONE` for build pack images
    `ALL` for standard images
- If `mount` option is specified (`All`) then pass through

This replaces the logic that use to be applied in create_job, but is now
skipped as `mount` always has an option set (model default).

Signed-off-by: David Caro <dcaro@wikimedia.org>