update_job - handle immutable fields
Bug: T409007

When attempting to patch an immutable field, kubernetes will reject
this, causing a http error to be returned to the user.

The `spec.selector` is immutable in a deployment, unfortunatly this
contains a number of labels, which are user-changeable.

Fallback to delete/create if the patch request returns a 422
(unprocessable), otherwise raise an error to the user.