k8s deployment object - reduce labels in selector
Bug: T409007

This field is immutable, meaning the job has to be deleted and
re-created for any change.

Currently we are carrying the labels used for features into the
selector, meaning any change to file logging, failure emails or mounts,
results in the deployment being impactful.

The selector labels is intended for k8s to match the pods to the
deployment, the jobs are unique by name, so use the minimum labels that
can be expected not to change.

Note: This is the same set as specified in the topology spread.