images: add digest only when passed
If the user did not pass a digest, then use the url without it.

This allows supporting the current flow of :
* You run a job without digest
* You rebuild the image
* You restart the job and it eventually pulls the new image

Otherwise it would have the digest hardcodded, and would never pull
the newer image.

Having it like this still allows for users/components-api to specify
the digest explicitly, and in a new build, update the job only with
the new digest if they want (users would have to use the API directly
or ` toolforge jobs load` though).

The right thing here will be to store the original request, and
refresh the digest on restart/update if the original request did not
pass a digest, but we can't do that yet.

Signed-off-by: David Caro <dcaro@wikimedia.org>