Move image resolution to core
Currently we are using the `Image` model to pass the user specified
image down to the runtime, where it is resolved into the real
`canonical_image`.

The `Image` object as constructed in `to_core_job` is not usable, as the
`container` is not fulfilled and thus the `job` is modified within the
runtime logic.

While this works and invalid images are bubbled up to the user due to
how the code is structured today, it makes other aspects such as job
diffs tricky.